// src/Api/models/User.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Child } from './Child';
export type User = {
    _id?: string;
    name?: string;
    email?: string;
    password?: string;
    phone?: string;
    children?: Array<Child>;
    isActive?: boolean;
};